dom/virtualfilesystem/test/mochitest/test_virtualfilesystem_event_callback.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1026350
-->
<head>
<meta charset="utf-8">
<title>Test Inputport Connection Event</title>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1187723">Test FileSystemProvider Event
</a>
<script type="application/javascript;version=1.8">

"use strict";

let Ci = SpecialPowers.Ci;
SimpleTest.waitForExplicitFinish();

function createRequest(aType, aOptions, aCallback) {
  let vfsService = SpecialPowers.Cc["@mozilla.org/tv/fakevirtualfilesystemservice;1"].getService(Ci.nsIVirtualFileSystemService);
  let requestManager = vfsService.getRequestManagerById(aOptions.fileSystemId);
  return requestManager.createRequest(aType, aOptions, aCallback);
}

function testEvent() {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

  return new Promise(function(resolve, reject) {
    let successCallback = {
      onSuccess: function(requestId, value, hasMore) {
        ok(true, "successCallback is invoked.");
      },
      onError: function(requestId, error) {}
    };
    let errorCallback = {
      onSuccess: function(requestId, value, hasMore) {},
      onError: function(requestId, error) {
        ok(true, "errorCallback is invoked.");
        resolve();
      }
    };
    let _fileSystemId = "dummyId";
    let _displayName = "dummyFileSystem";
    let _writable = true;
    let _openedFilesLimit = 10;
    let _requestId = 0;
    let _dirPath = "/dummy/path/";
    navigator.fileSystemProvider.addEventListener("readdirectoryrequested", function onReadDirectory(event){
      navigator.fileSystemProvider.removeEventListener("readdirectoryrequested", onReadDirectory);
      ok(true, "got event.");
      is(event.options.fileSystemId, _fileSystemId, "file system id must be the same");
      is(event.options.requestId, _requestId, "request id must be the same");
      is(event.options.directoryPath, _dirPath, "directory path must be the same");
      let data = {"isDirectory": false, "name": "test", "size": 0, "modificationTime": 0};
      event.successCallback(new Array(data), false);
      let abortoptions = SpecialPowers.Cc["@mozilla.org/virtualfilesystem/virtualfilesystem-abort-requested-options;1"].createInstance(Ci.nsIVirtualFileSystemAbortRequestedOptions);
        abortoptions.fileSystemId = _fileSystemId;
        abortoptions.requestId = _requestId;
        abortoptions.operationRequestId = _requestId;
        _requestId = createRequest(Ci.nsIVirtualFileSystemRequestManager.REQUEST_ABORT,
                                   abortoptions,
                                   errorCallback);
    });
    navigator.fileSystemProvider.addEventListener("abortrequested", function onabort(event){
      navigator.fileSystemProvider.removeEventListener("abortrequested", onabort);
      ok(true, "got event.");
      event.errorCallback("NotFound");
    });

    navigator.fileSystemProvider.mount({fileSystemId: _fileSystemId,
                              displayName: _displayName,
                              writable: _writable,
                              openedFilesLimit: _openedFilesLimit}).then(
      function() {
        ok(true, "mount ok.");

        let options = SpecialPowers.Cc["@mozilla.org/virtualfilesystem/virtualfilesystem-readdirectory-requested-options;1"].createInstance(Ci.nsIVirtualFileSystemReadDirectoryRequestedOptions);
        options.fileSystemId = _fileSystemId;
        options.requestId = _requestId;
        options.dirPath = _dirPath;
        _requestId = createRequest(Ci.nsIVirtualFileSystemRequestManager.REQUEST_READDIRECTORY,
                                   options,
                                   successCallback);
      },
      function(aError) {
        ok(false, "Fail to mount: " + aError);
        resolve();
      }
    );
  });
}


function runTest() {
  ok(navigator.fileSystemProvider, "should have navigator.fileSystemProvider");

  testEvent()
  .then(function() {
    info("test finished");
    SimpleTest.finish();
  });
}

SpecialPowers.pushPrefEnv({"set": [["device.storage.enabled", true],
                                    //to ignore app scope check.
                                   ["dom.ignore_webidl_scope_checks", true]]}, function() {
  SpecialPowers.pushPermissions(
    [{"type": "filesystemprovider", "allow": true, "context": document}], runTest);
});

 </script>
 </pre>
 </body>
 </html>
